refactor gt7car and fix unit test
